refactor(meshmap): code split
create utils.js and move some functions
