+++ plugins/lime-plugin-locate/src/style.less
@@ -23,3 +23,57 @@
   .leaflet-control-mapcentercoord-icon {
     opacity: 0
   }
+
+  .icon {
+    position: relative;
+    width:16px;
+    height:16px;
+    margin: 16px;
+    }
+    
+    .icon .name {
+    width: 32px;
+    position: absolute;
+    text-align: center;
+    top: 18px;
+    left: -8px;
+    }
+    
+    .icon-home:after{
+    border: 8px solid #000;
+    border-left-color: transparent;
+    border-right-color: transparent;
+    border-top-color: transparent;
+    height: 16px;
+    width: 0;
+    position: absolute;
+    top: -31px;
+    left: 0;
+     content: "";
+    }
+    .icon-home {
+    background-color: #000;
+    width: 15px;
+    height: 10px;
+    border-bottom-left-radius: 1px;
+    border-bottom-right-radius: 1px;
+    -webkit-border-bottom-left-radius: 1px;
+    -webkit-border-bottom-right-radius: 1px;
+    -moz-border-radius-bottomleft: 1px;
+    -moz-border-radius-bottomright: 1px;
+    position: relative;
+    bottom: 15px;
+    left: -18px;
+    }
+    .icon-home:before {
+    background-color: #000;
+    width: 2px;
+    height: 5px;
+    border-radius: 1px;
+    -webkit-border-radius: 1px;
+    -moz-border-radius: 1px;
+    position: absolute;
+    top: -6px;
+    left: 1px;
+    content: "";
+    }
